requirements.txt
# SunSim-IEC60904-Classifier Dependencies
# Core application
streamlit>=1.28.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Database (lazy-loaded, not required for basic functionality)
psycopg2-binary>=2.9.0

# Optional: Enhanced visualizations
plotly>=5.15.0

# Optional: PDF report generation
# reportlab>=4.0.0
